+++ src/main/java/com/github/patsalyukas/outsideclasses/FactoryForCards.java
@@ -1,9 +1,9 @@
 package com.github.patsalyukas.outsideclasses;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public interface FactoryForCards {
     Card createCard(String cardNumber, String expDate, String firstName, String lastName, int PIN, int CVI, String type);
 
-    Balance createBalance(Currency currency, BigInteger sum);
+    Balance createBalance(Currency currency, BigDecimal sum);
 }
