+++ src/test/java/com/github/patsalyukas/outsideclasses/DataBaseTest.java
@@ -2,7 +2,7 @@ package com.github.patsalyukas.outsideclasses;
 
 import org.junit.jupiter.api.Test;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -11,7 +11,7 @@ class DataBaseTest {
     DataBase dataBase = new DataBase(new BankCardFactory());
     Card card1 = new BankCard("4256123542131234", "12/21", "IVAN", "PETROV", 1532, 652, "DEBET");
     Card card2 = new BankCard("4256123542134526", "12/21", "IVAN", "PETROV", 1532, 652, "DEBET");
-    Balance balance = new Balance(Currency.RUB, new BigInteger("15000"));
+    Balance balance = new Balance(Currency.RUB, new BigDecimal("15000"));
 
     @Test
     void validateCard() {
